[Cellular Library] Correct memory estimation.

+++ common/src/cellular_pktio.c
@@ -136,7 +136,7 @@ static void _saveData( char * pLine,
     LogDebug( ( "_saveData : Save data %p with length %d", pLine, dataLen ) );
 
     pNew = ( CellularATCommandLine_t * ) Platform_Malloc( sizeof( CellularATCommandLine_t ) );
-    configASSERT( ( int32_t ) ( pNew != NULL ) );
+    configASSERT( ( pNew != NULL ) );
 
     /* Reuse the pktio buffer instead of allocate. */
     pNew->pLine = pLine;
@@ -254,7 +254,7 @@ static CellularATCommandResponse_t * _Cellular_AtResponseNew( void )
     CellularATCommandResponse_t * pNew = NULL;
 
     pNew = ( CellularATCommandResponse_t * ) Platform_Malloc( sizeof( CellularATCommandResponse_t ) );
-    configASSERT( ( int32_t ) ( pNew != NULL ) );
+    configASSERT( ( pNew != NULL ) );
 
     ( void ) memset( ( void * ) pNew, 0, sizeof( CellularATCommandResponse_t ) );
